\name{ForecastThis-package}
\alias{ForecastThis-package}
\alias{ForecastThis}
\docType{package}
\title{
Interface between R and the ForecastThis platform
}
\description{
Build models and generate predictions using the ForecastThis platform.
}
\details{
\tabular{ll}{
Package: \tab ForecastThis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-24\cr
License: \tab What license is it under?\cr
}

}
\author{
Liliana Medina

lmedina@365media.com
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
