% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict}
\alias{predict}
\title{Get Predictions using a ForecastThisDSX Predictive Model}
\usage{
predict(test.set = NULL, trained.model = NULL)
}
\arguments{
\item{test.set,}{data frame containing the test set for which we want to make predictions}

\item{trained.model,}{either the path to a jar model file downloaded from ForecastThisDSX or
an object of class \code{trained.model}.}
}
\value{
\code{test.results}, data frame with the predicted probabilities for each value of the target variable.
}
\description{
\code{predict} generates predictions for a test data set, using a ForecastThisDSX predictive model.
}

