% Generated by roxygen2 (4.0.2): do not edit by hand
\name{release}
\alias{release}
\title{Clear a ForecastThisDSX Model}
\usage{
release(trained.model = NULL)
}
\arguments{
\item{trained.model,}{an object of class \code{trained.model}}
}
\value{
\code{success}, boolean indicating whether object was successfully cleared
}
\description{
Clear a ForecastThisDSX Model
}

